requirements.txt
streamlit
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-readers-file
pandas